chore: 디렉토리 구조 개선
- src 디렉토리 추가
- 레시피 로직/UI 분리
- 파일명 케밥 케이스 적용
