--- projects/phase-01/soma0078FirstApp/app/(tabs)/recipes.tsx
@@ -1,5 +1,0 @@
-import { RecipesList } from "../components/recipes";
-
-export default function RecipesScreen() {
-  return <RecipesList />;
-}
