--- projects/phase-01/soma0078FirstApp/app/components/recipes/index.ts
@@ -1,1 +1,0 @@
-export * from "./RecipeList";
