+++ projects/phase-01/soma0078FirstApp/src/app/(tabs)/_layout.tsx
