+++ projects/phase-01/soma0078FirstApp/src/app/(tabs)/index.tsx
