+++ projects/phase-01/soma0078FirstApp/src/app/(tabs)/recipes.tsx
@@ -1,0 +1,5 @@
+import RecipeListScreen from "@/screens/recipes/recipe-list-screen";
+
+export default function RecipesScreen() {
+  return <RecipeListScreen />;
+}
