+++ projects/phase-01/soma0078FirstApp/src/app/_layout.tsx
@@ -1,4 +1,4 @@
-import "../global.css";
+import "../../global.css";
 import { Stack } from "expo-router";
 
 export default function RootLayout() {
